Removed sys.path.append(path/to/DaiPythonScorer.py
The reason I removed sys.path.append(path/to/DaiPythonScorer.py) was
because we already tell MiNiFi where to find the Python DaiPythonScorer.py
file. We do this through the 'Module Directory' property in the config.yml.
So, we don't need to include sys.path.append. We can run the code
'from DaiPythonScorer import *'