Removed Module Directory
I don't think the Module Directory property is working properly.
From reading on this property, it is suppose to use sys.path.append
before the script loads, so we can import our module in that location.
However, when I start minifi, it signals that it can't find the module.
When I directly write sys.path.append(path/to/modules) in the python
script, I don't receive the module not found error.