Updated Walkthrough for DAI Py Scorer & MiNiFi
I added some more dependencies for running the Driverless AI Python
Scoring Pipeline in the prerequisites. I also do mention the user
will need to download the Python Scoring Pipeline from Driverless AI.
In the last two sections, I added a forward slash to the start of the
path to the real-time and batch predictions data location. At the beginning
of the walkthrough I added a hyperlink to a YouTube video that shows
the user step by step how to run the deployment example. In the video,
I do let the user know they will have needed to download the Driverless
AI Python Scoring Pipeline to their local machine, but other than that,
I cover all the steps needed to run the deployment example.