Updates added based on nkpng2k's Feedback
In prerequisites section of mojo-nifi/ README, removed Java runtime since in the walkthrough
I go through the instruction to install it.

In prerequisites section narrowed it down to suggest user has a Driverless AI environment,
an Ubuntu 18.04 EC2 instance and cloned the dai-deployment-examples repo on their local
machine.

Instead of having the MOJO2 Runtime JAR uploaded to GitHub, in the README I walkthrough having
the user download the Driverless AI MOJO SP, which comes with the MOJO2 Runtime JAR.

Removed the uploaded mojo-pipeline/ folder that was downloaded from Driverless AI. It's no longer
on GitHub.

Removed the uploaded testData/ folder from GitHub. I include instructions on how to get the
data in the README. For batch scoring, the user uses example.csv that comes with mojo-pipeline/
folder. For real-time scoring, the user will run bash commands to break example.csv into smaller bits.

Updated the code in ExecuteDaiMojoScoringPipeline.java (custom NiFi processor)

- on line 168, I use Java's getLogger for proper logging instead of dumping stacks to stderr/stdout
- on line 236, I put the mojoPipelineUUID before the scoring occurs
- on line 255, RecordSetWriter is instantiated as a try resource and it has multilevel inheritance
to AutoCloseable, so I removed the try catch statements that were used to manually close the writer
since when execution leaves the try resource block, the writer is automatically closed.
- Originally the predict() method was doing more than just predicting, so I created a method
called convertRecordToMojoFrame() that translates the NiFi Record to a MojoFrame and I also
created another method called convertMojoFrameToRecord() that translates the MojoFrame to
a NiFi Record. So, now if you look at lines 238 to 245 or lines 265 to 270, you will see the
NiFi record is converted to a MojoFrame, the model is executed on the MojoFrame to do scoring,
and the scored MojoFrame is converted to a NiFi scored record.