Updated the code so now there is only 1 call to predict()
Before this update, the code had 2 different calls to the predict() method. The first predict() call was
performed on the first NiFi record. The reason for doing this was to get the write schema from the scored
NiFi record since NiFi doesn't know the write schema ahead of time. However, a better approach with the help
of the Driverless AI MOJO Pipeline model was to get the scored MojoFrame Meta Data from the model since it
knows the scored schema ahead of time. The development involved with adding this update required converting
from DAI MojoFrame Meta Data to NiFi Record Schema since NiFi doesn't understand DAI MojoFrame Meta Data. In
the long run, getting the MojoFrame Meta Data from the model would save of some execution time since it only
needs to be done once after we load the MojoPipeline model. Then later in the code we pass the scored
MojoRecordSchema to NiFi's writerFactory.getSchema(...) method to create the scored writeSchema.