Updated DAI MOJO Flink integration
Added to README.md the version of Driverless AI used to build the MOJO Scoring Pipeline and the
version of the MOJO Scoring Pipeline. Replaced places where nifi was written with Flink. Changed
Open custom TCP port to be 8081 instead of 8080.
Updated BatchPredHydCoolCond.java and RealTimePredHydCoolCond.java with a different file path to
pipeline.mojo that should be easier to reproduce.