Updated README instructions to refer Flink 1.10.0 for better compatibility with CSA
Updated the README instructions to refer to Flink 1.10.0 for better compatibility with Cloudera Streaming Analytics (CSA).
Updated some of the code in Flink Batch and Real-Time data pipeline for the directory path it pulls input data from.
Also updated the screenshots in the README, so Flink's version number doesn't appear. The changes in UI between Flink 1.11.1
and 1.10.0 do not appear to be different.