Updated links and DAI MOJO SP version in README
Updated outdated links for dai-deployment-examples repo from forked to h2oai one.

Updated NiFi download link to the archived one since the previous no longer is supported.

Updated test environment related info for DAI version to be 1.9.0 for that was the one that the MOJO SP was built on. Updated MOJO SP version to be 2.4.8 since that was the Java Runtime API version used in building this custom NiFi Processor ExecuteDaiMojoScoringPipeline.